+++ src/main/java/com/konrad/garagev3/model/dao/Role.java
@@ -1,11 +1,8 @@
 package com.konrad.garagev3.model.dao;
 
 import lombok.*;
-import org.springframework.stereotype.Component;
 
 import javax.persistence.*;
-import java.io.Serializable;
-import java.util.Objects;
 
 @Data
 @Builder
@@ -17,6 +14,6 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "role_id")
-    private int id;
+    private long id;
     private String role;
 }
