+++ src/main/java/com/konrad/garagev3/repository/RoleRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository("roleRepository")
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     Role findByRole(String role);
-    Role findById(int id);
+    Role findById(long id);
     List<Role> findAll();
 }
